又坏了

+++ App/Inc/motor.h
@@ -1,8 +1,8 @@
 #include "include.h"
 
-extern float32_t ftm_quad_values[3];
+extern int16 ftm_quad_values[3];
 extern float32_t filter_wight[3];
-extern uint16 motor_pulse;
+extern int16 motor_pulse;
 
 extern int8 motor_protection_switcher;
 extern int8 motor_out_of_order;
